Initial commit - Launching the CMS Airship - Public Beta
This is the result of months of development time. It's not finished,
but it's stable enough to make public.

Update readme, add license file

Update composer.json

Oops

Begin example config, work on documentation

Blogroll is a bad word

I think this is a sensible layout for the documentation

Default to 32 bytes of entropy, never 0

Add the proper rewrite rule to the Caddyfile.

Did I mention that Caddy is amazing?

Use strict mode for sessions

Early Voog/wysihtml integration

Incremental improvement.

Remove files