More randomized audits.
We're using peer verification on key updates against what the server claims, so we don't need to add the complexity for the package updates themselves. If we trust the public key, we explicitly trust the developer who controls the corresponding secret key. In most cases, Paragon Initiative Enterprises has no control over the secret key.