Read this commit message very carefully.
This was reported by @lukasreschke on HackerOne as a stored
XSS vulnerability. We thank Lukas for taking the time to report
this to us so we can make Airship more secure.

However, it only affects Microsoft Internet Explorer and
**only works because IE does something incredibly stupid**.

XSS bugs are a hard problem, and we have to draw a line in the
sand somewhere. We can only do so much to prevent rogue browser
developers' custom misbehaviors from harming their users from
the server side.

We alluded to this principle [here](https://paragonie.com/blog/2016/03/client-authenticity-is-not-server-s-problem).

The bug report in question is at https://hackerone.com/reports/151231